requirements.txt
awsebcli==3.12.1
boto3==1.5.6
botocore==1.8.20
cement==2.8.2
colorama==0.3.7
Django==2.0
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
imgurpython==1.1.7
jmespath==0.9.3
olefile==0.44
pathspec==0.5.0
Pillow==4.3.0
python-dateutil==2.6.1
python-memcached==1.59
pytz==2017.3
PyYAML==3.12
requests==2.9.1
s3transfer==0.1.12
semantic-version==2.5.0
six==1.11.0
tabulate==0.7.5
termcolor==1.1.0
virtualenv==15.1.0
virtualenvwrapper-win==1.2.5
websocket-client==0.44.0
